Merged cs and p2p model in one codebase
Added a launcher that uses a "DAGonFS.ini" configuration file to get all the input parameters of the commando to run to start DAGonFS execution.
Now, the entire repository incorporates the two versions of DAGonF (client-server and peer-to-peer). To run any of them, it's just needed to specify it in the configuration file.